refactor(github): refactor the github module of the data pipeline
- [x] implement modules for the github module contexts;
